+++ Forge/src/main/java/fuzs/resourcepackoverrides/client/gui/components/HoverableAbstractSelectionList.java
@@ -1,0 +1,8 @@
+package fuzs.resourcepackoverrides.client.gui.components;
+
+import net.minecraft.client.gui.widget.list.AbstractList;
+
+public interface HoverableAbstractSelectionList<E extends AbstractList.AbstractListEntry<E>> {
+
+    E resourcepackoverrides$getHovered();
+}